+++ src/Commands/InstallCommand.php
@@ -5,10 +5,12 @@ namespace Laravel\Octane\Commands;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
 use Laravel\Octane\Swoole\SwooleExtension;
+use Throwable;
 
 class InstallCommand extends Command
 {
-    use Concerns\InstallsRoadRunnerDependencies;
+    use Concerns\InstallsFrankenPhpDependencies,
+        Concerns\InstallsRoadRunnerDependencies;
 
     /**
      * The command's signature.
@@ -34,12 +36,13 @@ class InstallCommand extends Command
     {
         $server = $this->option('server') ?: $this->choice(
             'Which application server you would like to use?',
-            ['roadrunner', 'swoole'],
+            ['roadrunner', 'swoole', 'frankenphp'],
         );
 
         return (int) ! tap(match ($server) {
             'swoole' => $this->installSwooleServer(),
             'roadrunner' => $this->installRoadRunnerServer(),
+            'frankenphp' => $this->installFrankenPhpServer(),
             default => $this->invalidServer($server),
         }, function ($installed) use ($server) {
             if ($installed) {
@@ -111,6 +114,44 @@ class InstallCommand extends Command
     {
         if (! resolve(SwooleExtension::class)->isInstalled()) {
             $this->warn('The Swoole extension is missing.');
+        }
+
+        return true;
+    }
+
+    /**
+     * Install the FrankenPHP server.
+     *
+     * @return bool
+     */
+    public function installFrankenPhpServer()
+    {
+        if (! $this->confirm("FrankenPHP's Octane integration is in beta and should be used with caution in production. Do you wish to continue?")) {
+            return false;
+        }
+
+        $gitIgnorePath = base_path('.gitignore');
+
+        if (File::exists($gitIgnorePath)) {
+            $contents = File::get($gitIgnorePath);
+
+            $filesToAppend = collect(['/caddy', 'frankenphp', 'frankenphp-worker.php'])
+                ->filter(fn ($file) => ! str_contains($contents, $file.PHP_EOL))
+                ->implode(PHP_EOL);
+
+            if ($filesToAppend !== '') {
+                File::append($gitIgnorePath, PHP_EOL.$filesToAppend.PHP_EOL);
+            }
+        }
+
+        $this->ensureFrankenPhpWorkerIsInstalled();
+
+        try {
+            $this->ensureFrankenPhpBinaryIsInstalled();
+        } catch (Throwable $e) {
+            $this->error($e->getMessage());
+
+            return false;
         }
 
         return true;