+++ src/Commands/stubs/frankenphp-worker.php
@@ -1,0 +1,3 @@
+<?php
+
+require __DIR__.'/../vendor/laravel/octane/bin/frankenphp-worker.php';